* master: (470 commits)
  Better handling of tweepy errors.
  Truncating more of the body due to payload size limits.
  Accidentally marking stories in active feed as read when marking other feeds as read.
  New experimental data collector for debugging feeds over time.
  Fixing broken notifications for people with no notifications.
  Fixing a couple Twitter and YouTube fetch issues.
  iOS v7.0.0 testflight beta.
  Gotham is the new default font.
  Notifications should start at the top on every open.
  Fixing notification popover placement on ipad.
  Fixing crash when remvoing multiple notifications at once.
  Fixing all ios8 warnings. v7.0.0 is ready to go to TestFlight!
  Moving to UIAlertController. Got a bunch more.
  Upping redis timeout to handle possibility that users are still losing their tokens.
  Fixing warnings.
  Fixing styling on notifications titles.
  Fixing notifications on ipad. Notifications are also now updated after saving.
  Attempting to allow both old form and new form lists in parameters. looks[]=like&looks[]=this instead of looks=like&looks=this
  Feed ids is a list.
  Handling new style list parameters.
  ...

+++ clients/ios/Classes/ThemeManager.m
@@ -203,7 +203,7 @@ NSString * const ThemeStyleDark = @"dark";
         colors = [NSMutableSet set];
     }
     
-    [colors addObject:[NSString stringWithFormat:@"0x%06lX", rgbValue]];
+    [colors addObject:[NSString stringWithFormat:@"0x%06lX", (long)rgbValue]];
     
     NSLog(@"all unique colors: %@", [[colors allObjects] sortedArrayUsingSelector:@selector(compare:)]);  // log
 }
